+++ CHANGELOG.rst
@@ -25,7 +25,9 @@ Changed
 - feat: optimisations for MeanResponseTransformer, further optimisations for  MappingTransformer `#451 <https://github.com/azukds/tubular/issues/451>_`
 - feat: optimisation chnages to DatetimeSinusoidCalculator, added 'return_native_override' argument to DatetimeSinusoidCalculator, reduced with_columns being called many times. https://github.com/azukds/tubular/issues/465
 - chore: refactored unit weight handling into method of WeightColumnMixin, tightened up handling
-- feat: optimisation changes to AggregateRowOverColumnsTransformer, BaseAggregationTransformer
+- feat: optimisation changes to AggregateRowsOverColumnTransformer, BaseAggregationTransformer
+- feat: added AggregateColumnsOverRowTransformer `#451 <https://github.com/azukds/tubular/issues/385>_`
+- chore: renamed AggregateRowOverColumnsTransformer to AggregateRowsOverColumnTransformer
 - placeholder
 - placeholder